Menu
Fixed some errors in MenuController.
