+++ src/main/controllers/MenuController.java
@@ -1,14 +1,17 @@
 package main.controllers;
 
-import main.models.MenuModel;
+import main.models.Menu;
+import main.models.ObservableModel;
 import main.utils.CurrentScene;
+import main.views.View;
 
 public class MenuController extends Controller {
 
-    private MenuModel model;
+    private Menu model;
+    private View view;
 
 
-    public MenuController(MenuModel model) {
+    public MenuController(Menu model) {
         this.model = model ;
     }
 
@@ -16,4 +19,9 @@ public class MenuController extends Controller {
         this.model.setCurrentScene(currentScene);
     }
 
+    @Override
+    protected void setModel(ObservableModel model) { this.model = (Menu) model; }
+
+    @Override
+    protected void setView(View view) { this.view = view; }
 }
